[chroot] Do not use argh for parsing arguments
chroot helpers are expected to forward all arguments to
the binary that they launch. This does not work correctly
for `--` flags because the argh crate attempts to parse
them as arguments of chroot.

Do manual parsing of the argument list instead and correctly
forward arguments to the launched binary.

Change-Id: I4ca0b8556a6c804faa3c55ce5f876fb70572cee2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695363
Reviewed-by: Allan MacKinnon <allanmac@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>