[component_manager] Forward flags for storage open
When a storage capability is opened, it is always opened
with RW flags, ignoring the flags that were passed in by
the client.

This is problematic if the client passes in the DESCRIBE
flag (which happens when `dash` attempts to `cd` into
a storage directory under `ffx component explore`)

Instead of cloning the storage directory, this change
reopens it using the client provided flags and mode.

Also removed the StartReason propagation for storage
capability. The reason is always the same (component is
accessing a storage capability)

Test: fx test component_manager_tests
Change-Id: I5e9bd39cdf82c56132c716b671fd30e02a05b4ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694184
Reviewed-by: Gary Boone <gboone@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>