[settings] Convert light to Jobs
The light setting makes use of the custom change function feature of the
jobs framework, which allows using an arbitrary function to determine
whether or not to send an update to a listener rather than just using
equality. Custom change functions must be accompanied by an ID that is
used as part of the job signature, which is used by the jobs manager to
execute requests on a certain sequence and provide caches to store the
previous values for hanging gets.

This CL hashes the light group name to use as an ID so that subsequent
calls with the same light group name will hang until the value changes.

Fixed: 79044
Test: fx test setui_service_tests
      manually verified on device that light API works
Change-Id: I9d0aa1a73dbfcd715b9d8ced0671ee5222cca072
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682502
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: William Xiao <wxyz@google.com>
Reviewed-by: Sijie Chen <sijiec@google.com>