[debug_data] Flush to directory only after all VMOs are merged
Updates debug_data collection for v2 tests to only flush merged
profiles to the filesystem after all VMOs have been collected and
merged.
This prevents an OOM that occurs when some large chromium tests are
run in CFv2 instead of CFv1. The working theory for why OOMs occurred
is that the timing of debug data collection is slightly different in
CFv1 vs CFv2. Before this change, in CFv2 we may write to the directory
multiple times while a test suite is run, whereas in CFv1 writing to
the directory occurs once after the test completes.

The chromium test launches a few processes in parallel using the same
binary. Since each process creates its own VMO for reporting coverage
profile, given N parallel processes there are N VMOs active at once (if
we pretend each process makes only 1 VMO). In addition, the filesystem
used is memfs, so the copy in the filesystem is also in memory.

I suspect what occurs for v2 is that the first process' VMO is ready,
but the others are not. Creating the copy in the filesystem allocates
more memory, as well as a copy or two in memory to handle copying from
the VMO to an internal buffer, and then copying again to filesystem. In
the V2 case, these additional copies are made while the other (N-1)
active processes have active VMOs, which could push the system to OOM.
This doesn't occur in V1 because merging is completed before the write
to the filesystem, so there are fewer VMOs when the profile is written
to the filesystem.

Testing: update unittests
Run the chromium unit test on an emulator setup with 16GB memory, which
matches the emulator used in chromium's infra. Run the test once
without this change, verify the OOM occurs, then run the test with this
change and verify the OOM does not occur.

Bug: 101408
Bug: crbug.dev/1327387

Change-Id: I43af108e7029e1a33bfa8a9f1ac26c181c586c06
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691325
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>