[microbenchmarks] Run fuchsia_microbenchmarks as a component
Previously it was being run as a shell command for the main "perf test
mode" invocation.

Temporarily resurrect the runTestComponent() method for Components v1
(renamed to runTestComponentV1() here), which was previously removed
in https://fuchsia-review.googlesource.com/c/fuchsia/+/671240.

Change both runTestComponentV1() and runTestComponentV2() to take a
processRuns argument for doing multiple process invocations, in order
to preserve the current characteristics of this test.  The latter
method is not exercised with processRuns > 1 yet, but that will be
exercised when this test is converted to use Components v2.

Bug: 103374
Test: fx set terminal.x64 --release
  + fx test --e2e host_x64/microbenchmarks_test
Change-Id: I01813df88999119180a90aeb74db51771ddae51b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695107
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nathan Rogers <nathanrogers@google.com>
Fuchsia-Auto-Submit: Shai Barack <shayba@google.com>
Reviewed-by: Shai Barack <shayba@google.com>