[identity] Delete Token Management.
Fuchsia previously supported acquiring authentication and identity
tokens such as OAuth AccessTokens from service providers such as Google.
We don't have any products that currently use or intend to use this
platform capability and we already unhooked it from the external API
surface.

This CL deletes the remaining library implementation, the two
AuthProvider components we were maintaining, and the remaining vestiges
inside the account system that consumed configuration for auth
providers. If needed in the future we could resurrect the auth providers
and a simplified version of the token management library to provide a
per-system-account service.

Its a bit of a sad day since Satsuki and I poured a good amount of
effort into this code, but less code is less maintenance.

Bug: 101937,101939
Change-Id: I7ffc17b4b2c5285cf2392660653c6082937b633a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693908
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
API-Review: Shai Barack <shayba@google.com>