[nvme-cpp] New NVMe driver in C++
This CL is the first in a series of CLs that introduce a new, rewritten
NVMe driver to the tree.

This driver is written in C++ and uses more idiomatic code compared to
the older, C version (and it has tests!).

So far it does not support any I/O, but it sets up admin
submission/completion queues and runs IDENTIFY on the controller.

It also fills in some inspect information.

Test: fx test nvme-tests
Multiply: nvme-tests
Change-Id: Ifd8a42413e310e649712fed725998b20d5329f74
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686144
Reviewed-by: Christopher Anderson <cja@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>