[virtualization][vsock] Reacquire signals before polling the socket
The default async socket behaviour is to cache the read/write signal as
an optimization, allowing a client to read or write until failure.

The vsock device never wants to fail a read, and isn't able to fail a
write due to the credit system. To address this, we must reacquire
these signals before polling the socket to ensure freshness.

Note that this codepath is relatively rare -- a write signal is only
polled when credit hits zero without any RX, and a read signal is only
polled when the socket is empty and the connection needs to wait anyway.

Bug: 97355
Change-Id: I833e788c4f6ddcc577fe027b207663efcd18bbd2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695207
Commit-Queue: David Hastings <dahastin@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>