[diagnostics][sampler] Don't panic on underflow subtract
In Rust, unsigned underflow is unchecked in release mode but a
panic in debug mode. Sampler had been doing an underflow subtraction
and then checking if the result	was too	big. Now it does checked
subtraction.

This also changes a warning to be more informative.

Test: fx test //src/diagnostics/sampler
Fixed: 103584
Change-Id: I636c6fa987088a996d93d0bf834a9f513e6e0755
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695511
Commit-Queue: Chris Phoenix <cphoenix@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>