[starnix] Exit ffx starnix shell when remote task ends
The current code would not exit the shell because it waited for the
copy futures to end, which didn't happen until the user type something
on the terminal end generated an error trying to write to stdin.

This CL transforms the copy futures so that they never terminates, and
use try join so that the shell stops when the fuchsia component stops.

Change-Id: Idc97d7303fc63ce2d8f84303ac2e8d9afc3a6825
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695343
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>