[fdf][dfv2] Composite Device assembler
Create a class that will assemble composite devices in the way that
DFv1 expects them to be assembled.

This will help support device_add_composite in the compat shim.

Currently this class is not connected to anything, it will be
added to driver runner in a follow up CL.

Bug: 103306
TEST: composite-assembler-test
Change-Id: I30c69aba9c7c7f52fc8a81988be0cca5fde33daa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694808
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>