[amlogic] Adjust deadline parameters
The amlogic deadline parameters are too stringent which causes a lot of
thread context switching. Some of the tasks need more cpu time inorder
for them to accomplish in their deadline. The rationale for each change
is explained below:

vdec_irq: The HandleInterrupt was running out of CPU time which caused
it to have to be rescheduled. Generally deference is given to
interrupt routines and, so long as there operations are not
computational complex, they should be allowed to finish their
operations within one time slice.

h264_core: In steady state h264_core thread is only triggered when new
input is received and just pumps or reschedules the decoder. During
startup the first input packet received does have a cold startup period
so add a bit more CPU time to allow the thread to make progress without
getting descheduled.

Bug: 103225
Bug: b/235629262
Change-Id: I6e2170dba9a54f567855d1d2d6cab48e00310742
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695492
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Stefan Bossbaly <stefanbossbaly@google.com>