[wlan][sme] Add Rust channel indirection with MLME
Change `create_sme` in the SME library to take in a Rust channel instead
of an MLME FIDL channel. With this CL, SME and MLME will communicate
with a combination of a bidirectional FIDL channel and two
uni-directional Rust channels (used for MLME requests and MLME events).
This adds an extra layer of indirection for now, but later on when we
unify SME and MLME, the intermediate FIDL channel can be removed and
SME and MLME will communicate through the two Rust channels directly.

Bug: 66772
Test: Verify manually that connecting still works
Change-Id: Ib454cdb347854d5224a182299a3adf7907df5127
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694062
Fuchsia-Auto-Submit: Kiet Tran <kiettran@google.com>
Fuchsia-Auto-Submit: Dylan Swiggett <swiggett@google.com>
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>