[scenic][screen_capture2] implement Configure()
This CL works with the configure function of screen_capture2
which checks for argument existence and retreieves the buffer_count
from the ScreenCaptureBufferCollectionImporter.

Adds unittests that use the null_renderer as well as the
mock_renderer to ensure 100% testing. The mock_renderer is used to test
that if ImportBufferImage of the ScreenCaptureBufferCollectionImporter
fails, ReleaseBufferImage will be called the correct number of times.

Test: fx test screen_capture2_unittests
Change-Id: Ibeeaa3b4fe842d3455fcf5e17a49c6322cf2e897
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694825
Commit-Queue: Oliver Lance <ollylance@google.com>
Reviewed-by: Felipe Archondo <felipearchondo@google.com>