[virtualization] Extract the header from the command structs
We remove this field from the command structs to simplify reading
commands from the queue. Since all commands have a header and that
header contains the message type, we always read the header out of the
chain as part of command dispatch. By removing the header from these
structs the command implementations can just read the command specific
part of the request from the chain.

For now we keep the headers in the response structures since this is
something the command implementations will write entirely themselves.

Bug: 102870
Change-Id: Icaccad12b796d96bf251bc6f0cf50a1fc601682b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694544
Reviewed-by: David Hastings <dahastin@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>