[fidl][cpp] unknown interactions server side
Change Natural Types FIDL servers to be able to handle Unknown
Interactions.

For Open and Ajar protocols, add an additional base-class,
fidl::UnknownInteractionHandler, which receives information about which
kind of unknown interaction was received and allows the server to
respond to it.

Change dispatching to also accept an UnknownInteractionHandlerEntry
which enables unknown interaction handling. The
UnknownInteractionHandlerEntry tells the Dispatch function whether to
handle only one-way or both one-way and two-way interactions and
provides a function to use to send the unknown interaction reply for the
transport. It also provides the method to dispatch the protocol
implementation's handle_unknown_interacton method.

Bug: 88366
Test: fx test fidl_cpp_unknown_interactions_tests fidl_driver_transport_tests fidl_llcpp_unknown_interactions_tests

Change-Id: Ieef775421e8599ea71ff82dbe9a861a496dcabfa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686963
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Zachary Stewart <zstewart@google.com>
API-Review: Ian McKellar <ianloic@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>