[fx][rfc] Clarify "Area: FIDL" sections as samples
The RFC	generated contains "FIDL" area entries regardless of
actual area chosen. This can be	confusing. Expand the wording
of "Area: FIDL" headings to make it clearer that these are
sample and optional.

This is a continuation of fxr/693285 which was reverted by
fxr/695604. This CL updates a golden file that was affected but
presubmit tests didn't catch.

Test: fx test //tools/rfc
Fixed: 103195
Change-Id: I65a917d2eae5918b8d20ff17ef1edc1428bf29e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695686
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Commit-Queue: Chris Phoenix <cphoenix@google.com>