[fidldoc] Add a syscall generation template.
Adds a "syscall" template which can generate markdown for system calls.
This template is extremely simple: it basically copies the docstring
into a file with the system call name.

This will be used in a followup to generate the current syscall markdown
files from FIDL comments.

Change-Id: I933dce4975b175e89e45c7d4bfca461232da0e39
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695601
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Claudio Cherubino <ccherubino@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>