[virtualization] Implement GET_DISPLAY_INFO
GET_DISPLAY_INFO is a command that allows the driver to query how many
scanouts exist and what their geometry is (in pixels). Scanout here just
refers to a virtual display, which will eventually be view composited by
Flatland.

Since we have not yet laid the groundwork to integrate with Flatland or
create any views, we provide a very simple implementation that returns a
single scanout with hardcoded geometry. This will evolve as we built out
the real Flatland integration.

Bug: 102870
Test: fx test virtio_gpu_unittests
      fx test device_tests
Change-Id: I2e167bb7cd512e50760fcb876b1cbc7fa503954c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/692303
Reviewed-by: Alexander Radchenko <alerad@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>