[starnix][binder] Ease padding restriction on SG transactions
The length of a buffer does not need to be padded in the buffer data
structure sent to binder. What DOES need to be padded is the total
buffer size sent to binder, since the buffer start addresses need to be
aligned. This assumes that the length will be padded by binder instead.

Change-Id: Ic9e78a7b74a4b64c5598f7beab3f5196bfd4a9ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/696022
Commit-Queue: Adam Lesinski <adamlesinski@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>