[driver_runtime] Synchronize irq callbacks.
Irqs are only supported for synchronized dispatchers,
and may only be unbound on the dispatcher thread on which
it was bound.

Added an AsyncIrq class that wraps around the user's
async_irq_t. It is similar to AsyncWait, except for the
fact there may be multiple callbacks per AsyncIrq.

Bug: 99975

Change-Id: Iad645b1acf0f8b14b769d15cae074a276da4d93f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689116
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Simon Shields <simonshields@google.com>