[fdf][driver-index] Update device group implementation
Rewrite how device groups are handled in the driver
index so that it can handle multibinding in DFv2.

Driver index now maps a list of device groups to
device group nodes. If it matches a node, it'll
return the list of device groups that contain that
node.

Bug: 103206
Tests: driver-index-unittests
Change-Id: I97339be41ae9ca88fecb562ba481b18659e54a15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694924
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>