[netstack3] Don't include SLL in RS w/ unspecified src
Previously we incorrectly sent router solicitation messages with the
source link-layer option when the source address is unspecified which
is an RFC violation.

This change introduces an either type for `Serializer`s
(`EitherSerializer`) to the `packet` crate. We cannot impl `Serializer`
for `Either` in the `packet` crate because `Either` implements
`GrowBuffer` and `ShrinkBuffer` and there is a blanket impl of
`Serializer` for all types that implement both `GrowBuffer` and
`ShrinkBuffer`. Also note that `packet`'s `Either` type is documented
to only support buffers.

Change-Id: I5b0ad3f5d525ea9318da9a0fae01df56d598d44b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694937
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>