[lowpan_driver_common] Don't remove routes that are still in use.
If there are multiple addresses with the same subnet, then removing
one of them would cause the on-link subnet route to be removed,
making the other devices on the subnet unreachable. This change
adds a table to keep track of addresses being added and removed
from the interface so that the subnet is only removed when all
of the addresses with that prefix have been removed.

Change-Id: I94273266290f511aba6b15e7b5cd030b7d45b7b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689687
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Commit-Queue: Robert Quattlebaum <rquattle@google.com>