[outgoing_directory] Copy into `sys/component/cpp`
This CL copies the `//sdk/lib/sys/component/llcpp` library
into `//sdk/lib/sys/component/cpp`. This new directory will
be the new home for this library as decided in the API calibration
conducted at http://fxrev.dev/691874.

The reason for this move is that the library will support not only LLCPP
but also the new unified bindings. Indeed, it already does so with the
`AddProtocol` overload that accepts `fidl::Server<T>*` as its first
argument.

The old library is officially deprecated and will be kept there until
all OOT clients are migrated to the `cpp` one.

Bug: 103203
Change-Id: Ibbf3b357484af74e886ffda5406785d8e027d029
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/696984
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Yaneury Fermin <yaneury@google.com>