[package-directory] RootDir::new error variant for missing meta.far
RootDir clients may want to know if the error was caused by the
meta.far not being present in storage (e.g. pkg-cache activating cache
packages, in which case an absent cache package is not an error), and
without this new variant it is difficult for clients to know which
variant corresponds to a missing meta.far (and we weren't guaranteeing
it wouldn't change).

Change-Id: Ibbf25318a74b9e7297fd0684f6459f1830475f56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/697125
Reviewed-by: Sen Jiang <senj@google.com>
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>
Commit-Queue: Ben Keller <galbanum@google.com>