[fidldoc] Work on syscall documentation generation.
Adds support to fidldoc to update a directory in-place, leaving
non-generated files untouched. This allows generating syscall
documentation while leaving the hand-written index files in place.

Updates the sycall template to pass the configuration in and hooks up
"generator instructions" in the output files on how to regenerate them.

Adds a script to run fidldoc to generate the syscall documentation.
Since the syscall documentation does not exist in fidl yet, this is done
to the side of the existing update-docs-from-fidl script. When we switch
over the documentation source, the new script will overwrite the old
update-docs-from-fidl.

Change-Id: Ic2618a62d135090ccad4e7b35bcafeb61c29b60a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/696188
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Claudio Cherubino <ccherubino@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>