[blobfs][host] Option for exporting compressed blob versions.
Adds an option to provide a path where compressed versions of blobs are
placed, and adds an entry to exported json to provide resulting path
information if the compressed version is generated.

Adds a test to verify that this new method works and generates a valid
blob.

Change-Id: I7b8a983901ee9c1614e1261e614d54968f3a97a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/692602
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>