[blobfs] Move write buffer to WriteInfo
As all reading is done now using the pager path, there is no need for
the VMO to be accessible outside of the write path.

Remove unused target compression size as the new direction for streaming
writes is to use offline compression.

Remove initializer for write_info_ when opening existing blobs, which
should save a bit of memory (it was only required for the target
compression size).

Test: No functional change.
Change-Id: I7a2c590b34d15d21185587a0971921b58ebfac77
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695524
Fuchsia-Auto-Submit: Brandon Castellano <bcastell@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>