[cmx] Remove overnetstack-with-serial.cmx
I'd try to migrate it, but it's currently unused, so I don't know how to
test that I successfully migrate it.
The CF team will gladly help recreate this as cml if needed when the
time comes for this to be in use.

Change-Id: I54e53cecc4877390e9477e732e41318bb31dfe70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/697154
Reviewed-by: Casey Dahlin <sadmac@google.com>
Fuchsia-Auto-Submit: Shai Barack <shayba@google.com>
Commit-Queue: Shai Barack <shayba@google.com>