[bt][fastpair] Evict procedures after default deadline
The GFPS spec states that the shared secret in a Pairing Procedure must
be evicted after 10 seconds (if no update is received).

Implement Future for a Fast Pair Procedure - the Future resolves if the
shared secret deadline is met. Evict the procedure in this case as the
procedure is no longer valid.

Implement Drop for a Fast Pair Procedure - if it's evicted (and
therefore dropped), any outstanding pairing request will be rejected.

Bug: 99757
Test: fx test bt-fastpair-provider-tests

Change-Id: I7fab51d8a91db58b474527d01b6abe54da03446d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693931
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Zac Bowling <zbowling@google.com>