[build] Only exclude gnu linker symbols when is_gcc
These symbols are created when DSO's are linked by gcc
toolchains. They show up even when lld links shared
objects against these ld.gold linked shared objects
too. For that reason it wasn't enough to just exclude
these symbols when is_gcc because those libraries
may not be built with gcc but are linking against
a library linked with a gnu linker. Libraries are
linking against the stubs so removing these symbols
from the stubs is enough to ensure lld linked
shared objects do not pull in these symbols so we
can just use is_gcc in the abi check as well. It
is important to note that the abi check is checking
against the actual shared object and not the stub
which is why the is_gcc check must exist in
verify_public_symbols as well as zircon_toolchain.

Fixes: 99723
Change-Id: Id1c88d4094fa7ac6371ad222c26119851bd30916
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689843
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Alex Brachet <abrachet@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>