[kernel][coresight] Remove tracking of base address from RomTable
This information is not needed, and to store this is  awkward in the
MMIO hwreg context, as the provided hwreg IO provider already
encapsulates the base address. Consequentially, since we no longer have
access to the base address, we make the callback passed to
RomTable::Walk be a function of the offset of each found component.

Moreover, we move `max_offset` from being a parameter to RomTable's
constructor to one of Walk(), as it is not the state of the table itself
but rather of the aperture implicitly being passed to Walk().

Change-Id: I354ab269f3a8da69f54b1f2a3ef0903c82ab3886
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694687
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>