[governance] Rename "UI" area back to "Graphics".
This is a partial revert of fxrev.dev/662233. That CL renamed
'Graphics' to 'UI' as an area for the Glossary, but also renamed the
functional area for the API Council.

Consolidating the various UI-adjacent areas into a single 'UI' area
might be a reasonable thing to do, but that would require agreement
from the API Council.

Change-Id: I06829a181a4a6c11d0d5475549a0216b1bc26208
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/697957
Reviewed-by: Emircan Uysaler <emircan@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>