[flatland] Add copy fallback for screenshots on emulator
This CL adds a copy mode to screen capture that can be used for cases
where host-visible render targets are unsupported, i.e. emulator.

Using the clients sysmem token, we set constraints for a copy fallback.
Render target constraints are set on an AttachToken which might fail.
In the failure case, we register a readback buffer collection with
VkRenderer to copy the render target.

Test: Existing tests pass on emulator!
Fixed: 91950
Change-Id: I5380ff5cfdc08a3cf27d303e8a3752c96b1fed05
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/696324
Reviewed-by: Yilong Li <liyl@google.com>
Reviewed-by: David Worsham <dworsham@google.com>
Reviewed-by: Josh Gargus <jjosh@google.com>
Reviewed-by: Felipe Archondo <felipearchondo@google.com>
Commit-Queue: Emircan Uysaler <emircan@google.com>