[governance] Make FEC the owners of governance/.
Before this change, `//docs/contribute/governance` was owned by the
docs team, and only the `rfcs/` subdirectory was owned by the
FEC. This CL adds the FEC to the OWNERS file. This probably didn't
matter much in the past, but there are now multiple FEC members who
are not global approvers, so it's worth fixing this.

This also removes some members of the docs team from the OWNERS file,
in the spirit of least privilege. The tech writing leads should make
sure to loop in the FEC for any governance-related changes.


Change-Id: Ia8986be8f16ca6be15c6e4e42c4f426ada414bf2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698059
Commit-Queue: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>