[fxfs] Fix shutdown race in test
Automatic volume unmount happens in a separate async task.
The test verifies that the volume was removed from the
mounted_volumes map, but the removal from the map happens before
the volume is terminated asynchronously. This allows the test to
complete before the volume has been terminated.

As volume termination happens under the write lock, we can call
`VolumesDirectory::terminate()` at the end of the test to block
until the lock is released, ensuring the volume has been terminated.

Bug: 103735
Test: fx test fxfs-tests
Change-Id: I470844f6bf5b5a576ebec7ad9b381d2977da27d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/697782
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>