[ffx] Refactor emu to separate config from launch
As described in go/ffx-emu-config-design, before we can implement our
plan for configuration editing in the emulator launcher, it's necessary
to separate the process of building a Launch Configuration from using it
to perform the launch.

This CL adds a new function to the EmulatorEngine trait, stage(), which
performs what was the first section of the start() function, i.e.
building the Launch Configuration. The stage() function processes the
configuration, populating the engine's flag data. Rather than return it
directly, we also add a function build_emulator_cmd() to the trait, so
both the engines or the commands themselves can reproduce the Command as
needed. That Command is now passed in to the start() function
explicitly.

This refactor enables us to do see, use, and modify the Launch
Configuration directly from the ffx emu commands themselves, enabling us
to save the Launch Config to the instance directory, edit it on the fly,
and display it to the user on request.

Bug: 98268
Change-Id: I8dcc1f75688301389f53317639cef035becab60a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/697003
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>