[starnix] Remove redundant is_shutdown test from UnixSocket::read
If this condition is met, the following peek/read call will return MessageReadInfo::default() because the queue is empty, the EAGAIN return would be skipped because the socket is shut down, and MessageReadInfo::default() would be returned anyway. So the check is redundant.

Change-Id: I8e5df0753707db243f4aeaf6124e51a958de7f21
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698403
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Theodore Dubois <tbodt@google.com>