[audio][tools] wav_player resilience to partial frames
If a WAV file ends with an incomplete frame, then for that
final frame wav_player will play the bytes that ARE included,
and pad the rest of the frame with silence. Previously, the
utility sent a packet.payload_size that was not a multiple
of the frame size, which caused the renderer to disconnect.
'bytes_per_packet_' (which determines packet.payload_size)
is defined as a multiple of frame_size_, so this "round up"
will never exceed the bounds of our packet.

This issue was uncovered by a WAV file provided to us from
another team, during hardware validation.

Change-Id: I770086ed71a39a90aeada918302123791fad5af3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688721
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>