[virtualization][vsock] Add RX support to the RW state
This CL adds support for guest RX to the ReadWrite state without
actually wiring it in through the connection to be used by the device.

This CL also addresses an overflow bug in the credit system. Credit is
meant to be a running count which (unlike in C++) needs to be explicit
during the overflow in Rust.

Tested:
 - Data integrity tests
 - Overflow tests

Bug: 97355

Change-Id: I359fa55f41c87a27764af697cc072250334390e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695742
Commit-Queue: David Hastings <dahastin@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>